+++ pyproject.toml
@@ -5,7 +5,6 @@ description = ""
 authors = ["Yangyang Li <yangyang.li@northwestern.edu>"]
 license = "MIT"
 readme = "README.md"
-packages = [{include = "academic_chatgpt"}]
 
 [tool.poetry.dependencies]
 python = "^3.9"
@@ -18,6 +17,9 @@ mdtex2html = "^1.2.0"
 pre-commit = "^3.2.1"
 ruff = "^0.0.259"
 
+
+[tool.poetry.scripts]
+chataca = "academic_chatgpt.__main__:main"
 
 
 [tool.ruff]
