+++ src/academic_chatgpt/functional_crazy.py
@@ -1,32 +1,35 @@
-from crazy_functions.read_abstract import read_article_write_abstract
-from crazy_functions.parse_project_source_code import parse_project_itself
-from crazy_functions.parse_project_source_code import parse_a_python_project
-from crazy_functions.parse_project_source_code import parse_a_c_project_header_file
-from crazy_functions.higher_order_functional_template import (
-    higher_order_functional_template,
+from .crazy_functions.func_template import func_template
+
+from .crazy_functions.parse_source_code import (
+    parse_project,
+    parse_c_header,
+    parse_python_project,
 )
 
 
+from .crazy_functions.read_abstract import read_artical_write_abstract
+
+
 def get_crazy_functionals():
-    from crazy_functions.generate_func_comment import (
+    from .crazy_functions.generate_func_comment import (
         generate_comment_for_function_for_batch,
     )
 
     return {
         "[Experiment] Please parse and deconstruct this project itself": {
-            "Function": parse_project_itself
+            "Function": parse_project
         },
         "[Experiment] Parse the entire py project (input the project root path)": {
             "Color": "stop",  # button color
-            "Function": parse_a_python_project,
+            "Function": parse_python_project,
         },
         "[Experiment] Parse the entire C++ project (input the project root path)": {
             "Color": "stop",  # button color
-            "Function": parse_a_c_project_header_file,
+            "Function": parse_c_header,
         },
         "[Experiment] Read tex paper and write abstract (input the project root path)": {
             "Color": "stop",
-            "Function": read_article_write_abstract,
+            "Function": read_artical_write_abstract,
         },  # button color
         "[Experiment] Batch generate function comments (input the project root path)": {
             "Color": "stop",
@@ -34,6 +37,6 @@ def get_crazy_functionals():
         },  # button color
         "[Experiment] Experimental functional template": {
             "Color": "stop",
-            "Function": higher_order_functional_template,
+            "Function": func_template,
         },  # button color
     }
